# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plant Ride API",
    "description": "In-plant cab booking, pooling, fleet compliance and cost-centre billing for LHS.\n\n1. Call **POST /api/auth/login** with a demo account (password `Plant@123`):\n   `LHS-40218` (rider), `DRV-2281` (driver), `admin` (desk + admin), `desk` (desk).\n2. Click **Authorize** and paste the returned `token`.\n",
    "contact": {
      "name": "SRM ECO TECH",
      "email": "info@srmecotech.com"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/rider/sos": {
      "post": {
        "tags": [
          "3. Rider"
        ],
        "summary": "SOS: alerts the desk and plant security with the cab's live location",
        "operationId": "sos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiderSosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SosResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/rider/ride-options": {
      "post": {
        "tags": [
          "3. Rider"
        ],
        "summary": "Shuttle, shared and exclusive options with ETA and free seats",
        "operationId": "rideOptions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RideOptionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rider/bookings": {
      "get": {
        "tags": [
          "3. Rider"
        ],
        "summary": "My trips: scope = upcoming | past | cancelled",
        "operationId": "list",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "upcoming"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "3. Rider"
        ],
        "summary": "Book a ride. Send an Idempotency-Key so a retried tap never books twice",
        "operationId": "create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated UUID, one per booking attempt",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/rider/bookings/{id}/rating": {
      "post": {
        "tags": [
          "3. Rider"
        ],
        "summary": "Rate a completed ride (feeds the vendor SLA report)",
        "operationId": "rate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateBookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/rider/bookings/{id}/cancel": {
      "post": {
        "tags": [
          "3. Rider"
        ],
        "summary": "Cancel any time before boarding",
        "operationId": "cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelBookingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/driver/location": {
      "post": {
        "tags": [
          "4. Driver"
        ],
        "summary": "GPS ping from the driver app: live position, overspeed check, geofence arrival",
        "operationId": "location",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationPingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LocationAckDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rider/destinations": {
      "get": {
        "tags": [
          "3. Rider"
        ],
        "summary": "Stops reachable from a pickup stop, with the fastest route to each",
        "operationId": "destinations",
        "parameters": [
          {
            "name": "fromStopId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DestinationDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rider/bookings/{id}": {
      "get": {
        "tags": [
          "3. Rider"
        ],
        "summary": "One booking with vehicle, driver contact, OTP, live progress and charge",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/public/track/{token}": {
      "get": {
        "tags": [
          "8. Public tracking"
        ],
        "summary": "Live progress for a shared tracking link (no login, no personal data)",
        "operationId": "track",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingDto"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/billing/summary": {
      "get": {
        "tags": [
          "6. Billing (admin)"
        ],
        "summary": "Charges by cost centre, vendor bill vs GPS, and compliance watch for a month",
        "operationId": "summary",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "Billing month yyyy-MM; defaults to the current month",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BillingSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/billing/export": {
      "get": {
        "tags": [
          "6. Billing (admin)"
        ],
        "summary": "Ride-level CSV for the month-end ERP (SAP) import",
        "operationId": "export",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RiderSosRequest": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "integer",
            "format": "int64"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "note": {
            "type": "string",
            "maxLength": 300,
            "minLength": 0
          }
        }
      },
      "SosResultDto": {
        "type": "object",
        "properties": {
          "alertId": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RideOptionsRequest": {
        "type": "object",
        "properties": {
          "fromStopId": {
            "type": "integer",
            "format": "int64"
          },
          "toStopId": {
            "type": "integer",
            "format": "int64"
          },
          "seats": {
            "type": "integer",
            "format": "int32",
            "maximum": 12,
            "minimum": 1
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "fromStopId",
          "toStopId"
        ]
      },
      "RideOptionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "bookable": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "etaMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "seatsTaken": {
            "type": "integer",
            "format": "int32"
          },
          "seatsFree": {
            "type": "integer",
            "format": "int32"
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "vehicleCode": {
            "type": "string"
          },
          "requiresApproval": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "RideOptionsResponse": {
        "type": "object",
        "properties": {
          "route": {
            "$ref": "#/components/schemas/RouteRef"
          },
          "from": {
            "$ref": "#/components/schemas/StopRef"
          },
          "to": {
            "$ref": "#/components/schemas/StopRef"
          },
          "rideKm": {
            "type": "number"
          },
          "rideMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "occupancyCap": {
            "type": "integer",
            "format": "int32"
          },
          "costCentreCode": {
            "type": "string"
          },
          "costCentreName": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RideOptionDto"
            }
          }
        }
      },
      "RouteRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "StopRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateBookingRequest": {
        "type": "object",
        "properties": {
          "fromStopId": {
            "type": "integer",
            "format": "int64"
          },
          "toStopId": {
            "type": "integer",
            "format": "int64"
          },
          "rideType": {
            "type": "string",
            "enum": [
              "SHARED",
              "EXCLUSIVE"
            ]
          },
          "seats": {
            "type": "integer",
            "format": "int32",
            "maximum": 12,
            "minimum": 1
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "riderName": {
            "type": "string",
            "maxLength": 100,
            "minLength": 0
          },
          "riderPhone": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\+?[0-9 ()-]{7,20}$"
          },
          "riderEmail": {
            "type": "string",
            "format": "email",
            "maxLength": 120,
            "minLength": 0
          },
          "forGuest": {
            "type": "boolean"
          }
        },
        "required": [
          "fromStopId",
          "rideType",
          "riderName",
          "riderPhone",
          "seats",
          "toStopId"
        ]
      },
      "BookingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "bookingCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "PENDING_APPROVAL",
              "REQUESTED",
              "ASSIGNED",
              "ONBOARD",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW",
              "REJECTED"
            ]
          },
          "rideType": {
            "type": "string",
            "enum": [
              "SHARED",
              "EXCLUSIVE"
            ]
          },
          "seats": {
            "type": "integer",
            "format": "int32"
          },
          "forGuest": {
            "type": "boolean"
          },
          "riderName": {
            "type": "string"
          },
          "riderPhone": {
            "type": "string"
          },
          "riderEmail": {
            "type": "string"
          },
          "route": {
            "$ref": "#/components/schemas/RouteRef"
          },
          "fromStop": {
            "$ref": "#/components/schemas/StopRef"
          },
          "toStop": {
            "$ref": "#/components/schemas/StopRef"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          },
          "boardedAt": {
            "type": "string",
            "format": "date-time"
          },
          "droppedAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time"
          },
          "cancelReason": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          },
          "trackingToken": {
            "type": "string"
          },
          "trackingUrl": {
            "type": "string"
          },
          "vehicle": {
            "$ref": "#/components/schemas/VehicleRef"
          },
          "driver": {
            "$ref": "#/components/schemas/DriverContactDto"
          },
          "tripCode": {
            "type": "string"
          },
          "promisedEtaMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "promisedPickupAt": {
            "type": "string",
            "format": "date-time"
          },
          "progress": {
            "$ref": "#/components/schemas/RideProgressDto"
          },
          "costCentreCode": {
            "type": "string"
          },
          "costCentreName": {
            "type": "string"
          },
          "rideKm": {
            "type": "number"
          },
          "rideMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "distanceKm": {
            "type": "number"
          },
          "costAmount": {
            "type": "number"
          },
          "matchNote": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "ratingTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ratingNote": {
            "type": "string"
          },
          "canCancel": {
            "type": "boolean"
          },
          "canRate": {
            "type": "boolean"
          }
        }
      },
      "DriverContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
...
      "RateBookingRequest": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "ON_TIME",
                "CLEAN_CAB",
                "SAFE_DRIVING",
                "LONG_WAIT",
                "HARD_TO_FIND"
              ]
            },
            "maxItems": 5,
            "minItems": 0
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "rating"
        ]
      },
      "CancelBookingRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        }
      },
...
      "LocationPingRequest": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0
          },
          "speedKmh": {
            "type": "number",
            "format": "double",
            "maximum": 250.0
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "LocationAckDto": {
        "type": "object",
        "properties": {
          "overspeed": {
            "type": "boolean"
          },
          "speedLimitKmh": {
            "type": "integer",
            "format": "int32"
          },
          "autoArrivedAt": {
            "type": "string"
          }
        }
      },
...
      "DestinationDto": {
        "type": "object",
        "properties": {
          "stopId": {
            "type": "integer",
            "format": "int64"
          },
          "stopCode": {
            "type": "string"
          },
          "stopName": {
            "type": "string"
          },
          "routeId": {
            "type": "integer",
            "format": "int64"
          },
          "routeCode": {
            "type": "string"
          },
          "routeName": {
            "type": "string"
          },
          "rideMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "rideKm": {
            "type": "number"
          }
        }
      },
      "BookingSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "bookingCode": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "PENDING_APPROVAL",
              "REQUESTED",
              "ASSIGNED",
              "ONBOARD",
              "COMPLETED",
              "CANCELLED",
              "NO_SHOW",
              "REJECTED"
            ]
          },
          "rideType": {
            "type": "string",
            "enum": [
              "SHARED",
              "EXCLUSIVE"
            ]
          },
          "fromStopName": {
            "type": "string"
          },
          "toStopName": {
            "type": "string"
          },
          "when": {
            "type": "string",
            "format": "date-time"
          },
          "distanceKm": {
            "type": "number"
          },
          "vehicleCode": {
            "type": "string"
          },
          "costAmount": {
            "type": "number"
          },
          "cancelReason": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TrackingDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "stageLabel": {
            "type": "string"
          },
          "vehicleCode": {
            "type": "string"
          },
          "registrationNo": {
            "type": "string"
          },
          "routeLabel": {
            "type": "string"
          },
          "fromStopName": {
            "type": "string"
          },
          "toStopName": {
            "type": "string"
          },
          "vehicleAt": {
            "type": "string"
          },
          "etaToPickupMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "etaToDropMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProgressStopDto"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BillingSummaryDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string"
          },
          "draft": {
            "type": "boolean"
          },
          "sharingRule": {
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          },
          "costCentres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostCentreCharge"
            }
          },
          "vendors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VendorReconciliation"
            }
          },
          "compliance": {
            "$ref": "#/components/schemas/ComplianceWatch"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}